feat: add splash page
